FluchKarte und FluchkarteImpl in Kartentypen verändert dadurch FluchCards.java und FluchCardsImpl.java gelöscht
