--- app/src/main/java/com/example/munchkin/Karte/FluchCards.java
@@ -1,7 +1,0 @@
-package com.example.munchkin.Karte;
-
-public interface FluchCards  extends Karte{
-    void loseArmor(BodyPart bodyPart, int size);
-     void loseLevel(int lostLevels);
-
-}
