+++ app/src/main/java/com/example/munchkin/Karte/KartenTypen/FluchkarteImpl.java
@@ -1,5 +1,41 @@
 package com.example.munchkin.Karte.KartenTypen;
 
-public class FluchkarteImpl extends TürkarteImpl
-{
+import com.example.munchkin.Karte.BodyPart;
+
+public class FluchkarteImpl implements FluchKarte {
+    private int image;
+
+    //Übernommen von FluchCardsImpl
+    private int lostLevel;
+    private BodyPart bodyPart;
+
+//    Konstruktor aus FluchCardsImpl
+//    public FluchCardsImpl(int image, int lostLevel, BodyPart bodyPart){
+//        setImage(image);
+//        this.lostLevel=lostLevel;
+//        this.bodyPart=bodyPart;
+//
+//    }
+
+    @Override
+    public int getImage() {
+        return image;
+    }
+
+    @Override
+    public void setImage(int image) {
+        this.image=image;
+    }
+
+    //Übernommen von FluchCardsImpl
+
+    @Override
+    public void loseArmor(BodyPart bodyPart, int size) {
+
+    }
+
+    @Override
+    public void loseLevel(int currentLvl) {
+        currentLvl-= lostLevel;
+    }
 }
